requirements.txt
PySide6
llama-cpp-python==0.3.16
